requirements.txt
Flask
flask-login
python-nmap
requests
python-dotenv